third_party/blink/web_tests/wpt_internal/app-history/navigate-event/respondWith-appHistory-back.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // step_timeout because blink doesn't create history entries if onload has not
  // yet completed.
  window.onload = () => t.step_timeout(() => {
    appHistory.navigate("#foo").then(() => {
      assert_true(appHistory.canGoBack);
      appHistory.onnavigate = t.step_func(e => e.respondWith(Promise.resolve()));
      appHistory.back().then(t.step_func_done(() => {
        assert_equals(appHistory.entries().length, 2);
        assert_equals(appHistory.current, appHistory.entries()[0]);
      }));
    });
  }, 0);
}, "AppHistoryNavigateEvent.respondWith can intercept appHistory.back()");
</script>
